[skip ci] Use swift-driver on main scheme
The upstream recently started symlinking swift-driver to swift/swiftc
since https://github.com/apple/swift/commit/6a181fb04c3533a0c87559269ed9e7317d565d52
so the existence of swift-driver is now mandatory.

We will be able to use swift-driver after https://github.com/apple/swift-driver/pull/1523
is merged.

+++ tools/build/build-toolchain.sh
@@ -48,9 +48,12 @@ else
   echo "Using prebuilt cross compiler..."
   "$TOOLS_BUILD_PATH/install-base-toolchain" --scheme "$SCHEME"
   CROSS_COMPILER_DESTDIR=$PACKAGING_DIR/base-snapshot
-  # FIXME: Remove this once we fixed new swift-driver to use llvm-ar
-  # for WebAssembly target.
-  rm -f "$CROSS_COMPILER_DESTDIR/usr/bin/swift-driver"
+
+  # Only for release/5.9 and release/5.10
+  if [ "$SCHEME" = "release/5.9" ] || [ "$SCHEME" = "release/5.10" ]; then
+    # HACK: Remove swift-driver since the old swift-driver used "ar" instead of "llvm-ar"
+    rm -f "$CROSS_COMPILER_DESTDIR/usr/bin/swift-driver"
+  fi
 fi
 "$TOOLS_BUILD_PATH/build-llvm-tools" --toolchain "$CROSS_COMPILER_DESTDIR"
